Backend/requirements.txt
# ============================================================================
# Way2Rare Backend - Python Dependencies
# ============================================================================
# This file lists all Python packages needed for the backend API
# Install with: pip install -r requirements.txt
# ============================================================================

# FastAPI - Modern, fast web framework for building APIs with Python
# Provides automatic API documentation, request validation, and async support
fastapi==0.104.1

# Uvicorn - ASGI server to run FastAPI application
# The [standard] extra includes performance optimizations
uvicorn[standard]==0.24.0

# psycopg2-binary - PostgreSQL database adapter for Python
# The binary version includes pre-compiled binaries (easier installation)
# This allows us to connect to and query PostgreSQL databases
psycopg2-binary==2.9.9

# python-dotenv - Loads environment variables from .env file
# This lets us store database credentials and secrets in .env (not in code)
python-dotenv==1.0.0

# Pydantic - Data validation using Python type annotations
# FastAPI uses this to validate request/response data automatically
pydantic==2.5.0

# pydantic-settings - Settings management for Pydantic
# Helps manage application settings from environment variables
pydantic-settings==2.1.0

# python-multipart - Required for FastAPI to handle file uploads and form data
python-multipart==0.0.6

# boto3 - AWS SDK for Python
# Used for AWS services like Cognito (user authentication) and S3 (file storage)
boto3==1.29.7

# python-jose - JSON Web Token implementation
# Used for creating and verifying JWT tokens for authentication
# [cryptography] extra is needed for token signing/verification
python-jose[cryptography]==3.3.0

# passlib - Password hashing library
# [bcrypt] extra provides bcrypt hashing algorithm for secure password storage
passlib[bcrypt]==1.7.4


